finally removing the exodus methods file. Seems like a good cleanup so far.
