+++ src/Initialization.jl
@@ -1,3 +1,12 @@
+# TODO check these aren't outdated with older interface also add types to julia call
+function ex_get_cmap_params!(exoid::Cint, node_cmap_ids, node_cmap_node_cnts, elem_cmap_ids, elem_cmap_elem_cnts, processor)
+    error_code = ccall(
+        (:ex_get_cmap_params, libexodus), Cint,
+        (Cint, Ptr{void_int}, Ptr{void_int}, Ptr{void_int}, Ptr{void_int}, Cint),
+        exoid, node_cmap_ids, node_cmap_node_cnts, elem_cmap_ids, elem_cmap_elem_cnts, processor
+    )
+    exodus_error_check(error_code, "ex_get_cmap_params!")
+end
 
 # TODO add types
 """
@@ -39,6 +48,29 @@ function ex_get_init_info!(exoid::Cint, num_proc, num_proc_in_f, ftype)
                        (Cint, Ptr{Cint}, Ptr{Cint}, Ptr{UInt8}),
                         exoid, num_proc, num_proc_in_f, ftype)
     exodus_error_check(error_code, "ex_get_init_info!")
+end
+
+function ex_get_loadbal_param!(exoid::Cint,
+                               num_int_nodes, num_bor_nodes, num_ext_nodes,
+                               num_int_elems, num_bor_elems,
+                               num_node_cmaps, num_elem_cmaps,
+                               processor) # TODO get types right and sorted out
+    error_code = ccall(
+        (:ex_get_loadbal_param, libexodus), Cint,
+        (
+            Cint, 
+            Ptr{void_int}, Ptr{void_int}, Ptr{void_int}, 
+            Ptr{void_int}, Ptr{void_int}, 
+            Ptr{void_int}, Ptr{void_int}, 
+            Cint
+        ),
+        exoid, 
+        num_int_nodes, num_bor_nodes, num_ext_nodes, 
+        num_int_elems, num_bor_elems,
+        num_node_cmaps, num_elem_cmaps, 
+        processor
+    )
+    exodus_error_check(error_code, "ex_get_loadbal_param!")
 end
 
 # TODO should maybe make this parametric for Int32 vs. Int64
