+++ src/Maps.jl
@@ -1,4 +1,16 @@
 """
+    ex_get_map!(exoid::Cint, elem_map::Vector{T}) where {T <: ExoInt}
+"""
+function ex_get_map!(exoid::Cint, elem_map::Vector{T}) where {T <: ExoInt}
+    error_code = ccall(
+        (:ex_get_map, libexodus), Cint,
+        (Cint, Ptr{void_int}),
+        exoid, elem_map
+    )
+    exodus_error_check(error_code, "ex_get_map!")
+end
+
+"""
     read_element_map(exo::ExodusDatabase{M, I, B, F}, init::Initialization) where {M <: ExoInt, I <: ExoInt,
                                                                                    B <: ExoInt, F <: ExoFloat}
 """
