+++ src/NodeMaps.jl
@@ -1,3 +1,13 @@
+
+function ex_get_processor_node_maps!(exoid::Cint, node_mapi, node_mapb, node_mape, processor)
+    error_code = ccall(
+        (:ex_get_processor_node_maps, libexodus), Cint,
+        (Cint, Ptr{Cint}, Ptr{Cint}, Ptr{Cint}, Cint),
+        exoid, node_mapi, node_mapb, node_mape, processor
+    )
+    exodus_error_check(error_code, "ex_get_processor_node_maps")
+end
+
 # struct NodeMap <: FEMContainer
 #     internal_nodes::Vector{IntKind}
 #     border_nodes::Vector{IntKind}
